+++ src/hotel_reservations/utils.py
@@ -1,13 +1,18 @@
 import json
 import logging
 import os
+import re
 from typing import Any, Optional
 
+import numpy as np
+import pandas as pd
 import requests
 import yaml
 from databricks.sdk import WorkspaceClient
 from pyspark.ml.evaluation import RegressionEvaluator
-from pyspark.sql import DataFrame
+from pyspark.sql import DataFrame, SparkSession
+from pyspark.sql.functions import avg as mean
+from pyspark.sql.functions import stddev
 
 from hotel_reservations.types.project_config_types import ProjectConfig
 
@@ -115,3 +120,81 @@ def check_repo_info(repo_path: str, dbutils: Optional[Any] = None) -> tuple[str,
     db_repo_head_commit = repo_info.get("head_commit_id", "N/A")
 
     return db_repo_branch, db_repo_head_commit
+
+
+def generate_booking_ids_regex(existing_ids: list[str], num_new_ids=1000, prefix="INN") -> list[str]:
+    """
+    Generate a specified number of new booking IDs based on the given pattern using regex.
+
+    Args:
+        existing_ids (list): List of existing booking IDs (e.g., ['INN01205', 'INN01207']).
+        num_new_ids (int): Number of new booking IDs to generate.
+        prefix (str): The prefix for the booking IDs (default is 'INN').
+
+    Returns:
+        new_ids (list): A list of new booking IDs.
+    """
+    numeric_ids: list[int] = [
+        int(match.group(1)) for id in existing_ids if (match := re.search(rf"{prefix}(\d+)", id)) is not None
+    ]
+
+    max_id = max(numeric_ids)
+    padding_length = len(str(max_id))
+
+    new_ids = [f"{prefix}{str(i).zfill(padding_length)}" for i in range(max_id + 1, max_id + 1 + num_new_ids)]
+
+    return new_ids
+
+
+spark = SparkSession.builder.getOrCreate()
+
+
+def generate_synthetic_data(config: ProjectConfig, input_data: DataFrame, num_rows: int = 1000) -> DataFrame:
+    """Generates synthetic data in order to simulate data ingestion into the input data.
+
+    Args:
+        config (ProjectConfig): Project configuration file converted to dict, containing the catalog and schema where the data resides. Moreover, it contains the model parameters, numerical features, categorical features and the target variables.
+        input_data (DataFrame): Current input DataFrame with real data
+        num_rows (int, optional): Number of rows to add to the existing input data. Defaults to 1000.
+
+    Returns:
+        DataFrame: The added synthetic data, consisting of num_rows rows
+    """
+    num_rows = min(num_rows, 100000)  # Cap the number of rows
+    synthetic_data = {}
+
+    # Loop through numerical features with constraints
+    num_features = {key: {"min": feature.constraints.min} for key, feature in config.num_features.items()}
+
+    # Loop through the columns and generate data based on constraints
+    for col_name, constraints in num_features.items():
+        mean_val, std_val = input_data.select(mean(col_name), stddev(col_name)).first()
+
+        # Generate data and apply constraints
+        synthetic_data[col_name] = np.round(np.random.normal(mean_val, std_val, num_rows))
+
+        # Apply min constraints
+        synthetic_data[col_name] = np.maximum(synthetic_data[col_name], constraints["min"])
+
+    # Loop through categorical features with allowed values
+    cat_features = {
+        key: [
+            int(value) if isinstance(value, str) and value.isdigit() else value
+            for value in feature.encoding or feature.allowed_values or []
+        ]
+        for key, feature in config.cat_features.items()
+    }
+    for col_name, allowed_values in cat_features.items():
+        synthetic_data[col_name] = np.random.choice(allowed_values, num_rows)
+
+    # Create target variable (booking_status) as a random value
+    synthetic_data[config.target] = np.random.choice(["Not_Canceled", "Canceled"], num_rows, p=[0.9, 0.1])
+
+    existing_ids = input_data.select(config.primary_key).rdd.flatMap(lambda x: x).collect()
+    synthetic_ids = generate_booking_ids_regex(existing_ids, num_new_ids=num_rows)
+    synthetic_data[config.primary_key] = synthetic_ids
+
+    # Convert the synthetic data dictionary to a DataFrame
+    synthetic_df = spark.createDataFrame(pd.DataFrame(synthetic_data))
+
+    return synthetic_df